+++ whatever/package.json
@@ -13,6 +13,7 @@
     "dotenv": "^16.4.5",
     "framer-motion": "^11.0.24",
     "next": "14.1.0",
+    "next-redux-wrapper": "^8.1.0",
     "react": "^18",
     "react-dom": "^18",
     "react-kakao-maps-sdk": "^1.1.26",
