+++ whatever/src/redux/reduxStorage.ts
@@ -1,0 +1,17 @@
+import createWebStorage from "redux-persist/lib/storage/createWebStorage";
+
+const createNoopStorage = () => ({
+    getItem() {
+      return Promise.resolve(null);
+    },
+    setItem(_key: string, value: any) {
+      return Promise.resolve(value);
+    },
+    removeItem() {
+      return Promise.resolve();
+    },
+  });
+
+const storage = typeof window !== "undefined" ? createWebStorage("local") : createNoopStorage();
+
+export default storage;
